+++ bin/id.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:args/args.dart';
 import 'package:inno_bundle/utils/constants.dart';
 import 'package:uuid/uuid.dart';
@@ -6,11 +8,22 @@ const uuid = Uuid();
 
 /// Run to generate an App ID (as GUID)
 void main(List<String> arguments) {
-  print(START_MESSAGE);
+  final parser = ArgParser()
+    ..addOption('ns', help: "Namespace, ex: google.com")
+    ..addFlag('hf', defaultsTo: true, help: 'Print header and footer')
+    ..addFlag('help', abbr: 'h', negatable: false, help: 'Print help and exit');
 
-  final parser = ArgParser()..addOption('ns');
   final parsedArgs = parser.parse(arguments);
   final ns = parsedArgs['ns'] as String?;
+  final hf = parsedArgs['hf'] as bool;
+  final help = parsedArgs['help'] as bool;
+
+  if (hf) print(START_MESSAGE);
+
+  if (help) {
+    print("${parser.usage}\n");
+    exit(0);
+  }
 
   if (ns != null) {
     print(uuid.v5(Uuid.NAMESPACE_URL, ns));
@@ -18,5 +31,5 @@ void main(List<String> arguments) {
     print(uuid.v1());
   }
 
-  print(GUID_END_MESSAGE);
+  if (hf) print(GUID_END_MESSAGE);
 }